+++ src/Compilers/CSharp/Test/Syntax/Parsing/UsingDirectiveParsingTests.cs
@@ -5258,266 +5258,4 @@ using X = __makeref;
         }
         EOF();
     }
-
-    [Fact, WorkItem("https://github.com/dotnet/roslyn/issues/67989")]
-    public void AliasToPointerReferencedInExpression()
-    {
-        var text = """
-            using unsafe Alias = int*;
-
-            namespace N
-            {
-                internal unsafe class C
-                {
-                    const int Factor = 3;
-
-                    void M()
-                    {
-                        _ = Alias * Factor;
-                    }
-                }
-            }
-            """;
-        UsingTree(text);
-        CreateCompilation(text, options: TestOptions.UnsafeDebugDll).VerifyDiagnostics(
-            // (11,17): error CS0119: 'int*' is a type, which is not valid in the given context
-            //             _ = Alias * Factor;
-            Diagnostic(ErrorCode.ERR_BadSKunknown, "Alias").WithArguments("int*", "type").WithLocation(11, 17));
-
-        N(SyntaxKind.CompilationUnit);
-        {
-            N(SyntaxKind.UsingDirective);
-            {
-                N(SyntaxKind.UsingKeyword);
-                N(SyntaxKind.UnsafeKeyword);
-                N(SyntaxKind.NameEquals);
-                {
-                    N(SyntaxKind.IdentifierName);
-                    {
-                        N(SyntaxKind.IdentifierToken, "Alias");
-                    }
-                    N(SyntaxKind.EqualsToken);
-                }
-                N(SyntaxKind.PointerType);
-                {
-                    N(SyntaxKind.PredefinedType);
-                    {
-                        N(SyntaxKind.IntKeyword);
-                    }
-                    N(SyntaxKind.AsteriskToken);
-                }
-                N(SyntaxKind.SemicolonToken);
-            }
-            N(SyntaxKind.NamespaceDeclaration);
-            {
-                N(SyntaxKind.NamespaceKeyword);
-                N(SyntaxKind.IdentifierName);
-                {
-                    N(SyntaxKind.IdentifierToken, "N");
-                }
-                N(SyntaxKind.OpenBraceToken);
-                N(SyntaxKind.ClassDeclaration);
-                {
-                    N(SyntaxKind.InternalKeyword);
-                    N(SyntaxKind.UnsafeKeyword);
-                    N(SyntaxKind.ClassKeyword);
-                    N(SyntaxKind.IdentifierToken, "C");
-                    N(SyntaxKind.OpenBraceToken);
-                    N(SyntaxKind.FieldDeclaration);
-                    {
-                        N(SyntaxKind.ConstKeyword);
-                        N(SyntaxKind.VariableDeclaration);
-                        {
-                            N(SyntaxKind.PredefinedType);
-                            {
-                                N(SyntaxKind.IntKeyword);
-                            }
-                            N(SyntaxKind.VariableDeclarator);
-                            {
-                                N(SyntaxKind.IdentifierToken, "Factor");
-                                N(SyntaxKind.EqualsValueClause);
-                                {
-                                    N(SyntaxKind.EqualsToken);
-                                    N(SyntaxKind.NumericLiteralExpression);
-                                    {
-                                        N(SyntaxKind.NumericLiteralToken, "3");
-                                    }
-                                }
-                            }
-                        }
-                        N(SyntaxKind.SemicolonToken);
-                    }
-                    N(SyntaxKind.MethodDeclaration);
-                    {
-                        N(SyntaxKind.PredefinedType);
-                        {
-                            N(SyntaxKind.VoidKeyword);
-                        }
-                        N(SyntaxKind.IdentifierToken, "M");
-                        N(SyntaxKind.ParameterList);
-                        {
-                            N(SyntaxKind.OpenParenToken);
-                            N(SyntaxKind.CloseParenToken);
-                        }
-                        N(SyntaxKind.Block);
-                        {
-                            N(SyntaxKind.OpenBraceToken);
-                            N(SyntaxKind.ExpressionStatement);
-                            {
-                                N(SyntaxKind.SimpleAssignmentExpression);
-                                {
-                                    N(SyntaxKind.IdentifierName);
-                                    {
-                                        N(SyntaxKind.IdentifierToken, "_");
-                                    }
-                                    N(SyntaxKind.EqualsToken);
-                                    N(SyntaxKind.MultiplyExpression);
-                                    {
-                                        N(SyntaxKind.IdentifierName);
-                                        {
-                                            N(SyntaxKind.IdentifierToken, "Alias");
-                                        }
-                                        N(SyntaxKind.AsteriskToken);
-                                        N(SyntaxKind.IdentifierName);
-                                        {
-                                            N(SyntaxKind.IdentifierToken, "Factor");
-                                        }
-                                    }
-                                }
-                                N(SyntaxKind.SemicolonToken);
-                            }
-                            N(SyntaxKind.CloseBraceToken);
-                        }
-                    }
-                    N(SyntaxKind.CloseBraceToken);
-                }
-                N(SyntaxKind.CloseBraceToken);
-            }
-            N(SyntaxKind.EndOfFileToken);
-        }
-        EOF();
-    }
-
-    [Fact, WorkItem("https://github.com/dotnet/roslyn/issues/67989")]
-    public void AliasToTupleReferencedInExpression()
-    {
-        var text = """
-            using Alias = (int, int);
-
-            internal class C
-            {
-                void M()
-                {
-                    Alias.Equals(null, null);
-                }
-            }
-            """;
-
-        UsingTree(text);
-        CreateCompilation(text).VerifyDiagnostics();
-
-        N(SyntaxKind.CompilationUnit);
-        {
-            N(SyntaxKind.UsingDirective);
-            {
-                N(SyntaxKind.UsingKeyword);
-                N(SyntaxKind.NameEquals);
-                {
-                    N(SyntaxKind.IdentifierName);
-                    {
-                        N(SyntaxKind.IdentifierToken, "Alias");
-                    }
-                    N(SyntaxKind.EqualsToken);
-                }
-                N(SyntaxKind.TupleType);
-                {
-                    N(SyntaxKind.OpenParenToken);
-                    N(SyntaxKind.TupleElement);
-                    {
-                        N(SyntaxKind.PredefinedType);
-                        {
-                            N(SyntaxKind.IntKeyword);
-                        }
-                    }
-                    N(SyntaxKind.CommaToken);
-                    N(SyntaxKind.TupleElement);
-                    {
-                        N(SyntaxKind.PredefinedType);
-                        {
-                            N(SyntaxKind.IntKeyword);
-                        }
-                    }
-                    N(SyntaxKind.CloseParenToken);
-                }
-                N(SyntaxKind.SemicolonToken);
-            }
-            N(SyntaxKind.ClassDeclaration);
-            {
-                N(SyntaxKind.InternalKeyword);
-                N(SyntaxKind.ClassKeyword);
-                N(SyntaxKind.IdentifierToken, "C");
-                N(SyntaxKind.OpenBraceToken);
-                N(SyntaxKind.MethodDeclaration);
-                {
-                    N(SyntaxKind.PredefinedType);
-                    {
-                        N(SyntaxKind.VoidKeyword);
-                    }
-                    N(SyntaxKind.IdentifierToken, "M");
-                    N(SyntaxKind.ParameterList);
-                    {
-                        N(SyntaxKind.OpenParenToken);
-                        N(SyntaxKind.CloseParenToken);
-                    }
-                    N(SyntaxKind.Block);
-                    {
-                        N(SyntaxKind.OpenBraceToken);
-                        N(SyntaxKind.ExpressionStatement);
-                        {
-                            N(SyntaxKind.InvocationExpression);
-                            {
-                                N(SyntaxKind.SimpleMemberAccessExpression);
-                                {
-                                    N(SyntaxKind.IdentifierName);
-                                    {
-                                        N(SyntaxKind.IdentifierToken, "Alias");
-                                    }
-                                    N(SyntaxKind.DotToken);
-                                    N(SyntaxKind.IdentifierName);
-                                    {
-                                        N(SyntaxKind.IdentifierToken, "Equals");
-                                    }
-                                }
-                                N(SyntaxKind.ArgumentList);
-                                {
-                                    N(SyntaxKind.OpenParenToken);
-                                    N(SyntaxKind.Argument);
-                                    {
-                                        N(SyntaxKind.NullLiteralExpression);
-                                        {
-                                            N(SyntaxKind.NullKeyword);
-                                        }
-                                    }
-                                    N(SyntaxKind.CommaToken);
-                                    N(SyntaxKind.Argument);
-                                    {
-                                        N(SyntaxKind.NullLiteralExpression);
-                                        {
-                                            N(SyntaxKind.NullKeyword);
-                                        }
-                                    }
-                                    N(SyntaxKind.CloseParenToken);
-                                }
-                            }
-                            N(SyntaxKind.SemicolonToken);
-                        }
-                        N(SyntaxKind.CloseBraceToken);
-                    }
-                }
-                N(SyntaxKind.CloseBraceToken);
-            }
-            N(SyntaxKind.EndOfFileToken);
-        }
-        EOF();
-    }
 }
